+++ src/model/easing/Expo.java
@@ -2,19 +2,19 @@ package model.easing;
  
 public class Expo extends Easing {
 	
-	public  float  easeIn(float t,float b , float c, float d) {
-		return (t==0) ? b : c * (float)Math.pow(2, 10 * (t/d - 1)) + b;
+	public  double  easeIn(double t,double b , double c, double d) {
+		return (t==0) ? b : c * (double)Math.pow(2, 10 * (t/d - 1)) + b;
 	}
 	
-	public  float  easeOut(float t,float b , float c, float d) {
-		return (t==d) ? b+c : c * (-(float)Math.pow(2, -10 * t/d) + 1) + b;	
+	public  double  easeOut(double t,double b , double c, double d) {
+		return (t==d) ? b+c : c * (-(double)Math.pow(2, -10 * t/d) + 1) + b;	
 	}
 	
-	public  float  easeInOut(float t,float b , float c, float d) {
+	public  double  easeInOut(double t,double b , double c, double d) {
 		if (t==0) return b;
 		if (t==d) return b+c;
-		if ((t/=d/2) < 1) return c/2 * (float)Math.pow(2, 10 * (t - 1)) + b;
-		return c/2 * (-(float)Math.pow(2, -10 * --t) + 2) + b;
+		if ((t/=d/2) < 1) return c/2 * (double)Math.pow(2, 10 * (t - 1)) + b;
+		return c/2 * (-(double)Math.pow(2, -10 * --t) + 2) + b;
 	}
 
 }
